Add mimetype for .gn files in the context of presubmits.
".gn" was not a recognized extension for the `mimetypes.guess_type` function.  As a result, files like "BUILD.gn" weren't added to affected files, and so some presubmits fail.  (Specifically, the presubmit which requires new header files to be added to a BUILD.gn file.)

Bug: b:456124666
Change-Id: I38c4d8007e242acd218cf130d3b81d71b0c189ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/7102959
Reviewed-by: Gavin Mak <gavinmak@google.com>
Commit-Queue: Terrence Reilly <treilly@google.com>